[chromium] Don't call glDeleteTexture(0) in TextureManager https://bugs.webkit.org/show_bug.cgi?id=66862
Delete(0) is allowed in standard OpenGL, but not in Chrome.
See http://code.google.com/p/chromium/issues/detail?id=85268

Patch by Iain Merrick <husky@google.com> on 2011-08-24
Reviewed by James Robinson.

* platform/graphics/chromium/TextureManager.cpp:
(WebCore::TextureManager::deleteEvictedTextures):

Canonical link: https://commits.webkit.org/82693@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@93743 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebCore/platform/graphics/chromium/TextureManager.cpp
@@ -134,8 +134,10 @@ void TextureManager::addTexture(TextureToken token, TextureInfo info)
 void TextureManager::deleteEvictedTextures(GraphicsContext3D* context)
 {
     ASSERT(context == m_associatedContextDebugOnly);
-    for (size_t i = 0; i < m_evictedTextureIds.size(); ++i)
-        GLC(context, context->deleteTexture(m_evictedTextureIds[i]));
+    for (size_t i = 0; i < m_evictedTextureIds.size(); ++i) {
+        if (m_evictedTextureIds[i])
+            GLC(context, context->deleteTexture(m_evictedTextureIds[i]));
+    }
     m_evictedTextureIds.clear();
 }
